api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
neo4j==5.15.0
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
PyYAML==6.0.1
pdfplumber==0.9.0
python-docx==1.1.0
requests==2.31.0
